Move cell facet param functions into more cohesive module
